feat: 날짜, 시간 포맷 로직 util 함수로 분리
