+++ src/entities/ui/information/index.tsx
@@ -1,5 +1,6 @@
 import ChipInfo from '@/shared/common/ui/chip-info';
 import OpenBadge from '@/shared/common/ui/open-badge';
+import { formatDate, formatTime } from '@/shared/lib/utils';
 import { Progress } from '@/shared/ui/progress';
 import Image from 'next/image';
 
@@ -41,16 +42,9 @@ export default function Information({
   gatheringsInfo,
   participantList,
 }: InformationProps) {
-  // 날짜, 시간 계산
-  const eventDate = new Date(gatheringsInfo.dateTime);
-
-  const month = eventDate.getMonth() + 1;
-  const date = eventDate.getDate();
-  const hour = eventDate.getHours();
-  const minute = eventDate.getMinutes();
-
-  const formattedDate = `${month}월 ${date}일`;
-  const formattedHour = `${hour}:${minute}`;
+  // 날짜, 시간
+  const formattedDate = formatDate(gatheringsInfo.dateTime);
+  const formattedTime = formatTime(gatheringsInfo.dateTime);
 
   // 텍스트 색상
   const textColor =
@@ -76,7 +70,7 @@ export default function Information({
           </div>
           <div className="mt-3 space-x-2">
             <ChipInfo type="date">{formattedDate}</ChipInfo>
-            <ChipInfo type="time">{formattedHour}</ChipInfo>
+            <ChipInfo type="time">{formattedTime}</ChipInfo>
           </div>
         </div>
         {/* bottom */}
